refactor: #365 WriteQuestBaseVC 상속으로 리팩토링
