+++ ByeBoo-iOS/ByeBoo-iOS/Presentation/Feature/Quest/ViewController/WriteActiveTypeQuestViewController.swift
@@ -11,9 +11,8 @@ import UIKit
 import Kingfisher
 import Mixpanel
 
-final class WriteActiveTypeQuestViewController: BaseViewController {
-    
-    private let rootView = WriteActiveTypeQuestView()
+final class WriteActiveTypeQuestViewController: WriteQuestBaseViewController<WriteActiveTypeQuestView> {
+
     private let viewModel: WriteActiveTypeViewModel
     private var cancellables = Set<AnyCancellable>()
     var questMode: QuestMode = .write
@@ -25,55 +24,24 @@ final class WriteActiveTypeQuestViewController: BaseViewController {
     private var answerText: String = ""
     private var emotionState: String = ""
     private var image: UIImage = UIImage()
-    private var isKeyboardUsed: Bool = false
-    private var keyboardFrameInWindow: CGRect = .zero
-    private var currentKeyboardOffset: CGFloat = 0
-    private var previousTextViewHeight: CGFloat = 0
+
     private var isImageChanged: Bool = false
     private var originalImageKey: String = ""
     
     private let bottomSheetViewController = EmotionBottomSheetViewController()
     
-    override func loadView() {
-        view = rootView
-    }
-    
     init(viewModel: WriteActiveTypeViewModel){
         self.viewModel = viewModel
-        super.init(nibName: nil, bundle: nil)
+        super.init(rootView: WriteActiveTypeQuestView())
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
-    override func viewWillAppear(_ animated: Bool) {
-        NotificationCenter.default.addObserver(
-            self,
-            selector: #selector(textViewMoveUp),
-            name: UIResponder.keyboardWillShowNotification, object: nil
-        )
-        NotificationCenter.default.addObserver(
-            self,
-            selector: #selector(textViewMoveDown),
-            name: UIResponder.keyboardWillHideNotification, object: nil
-        )
-        isKeyboardUsed = false
-    }
-    
     override func viewDidLoad() {
         super.viewDidLoad()
-        ByeBooNavigationBar.makeNavigationBar(
-            navigationItem: self.navigationItem,
-            navigationController: self.navigationController,
-            type: .confirmAndBack("완료", header: .clear),
-            action: #selector(back),
-            secondAction: #selector(confirmButtonDidTap)
-        )
         
-        self.navigationItem.rightBarButtonItem?.isEnabled = false
-        
-        setGesture()
         setDelegate()
         bind()
         presentPhotoPicker()
@@ -93,67 +61,21 @@ final class WriteActiveTypeQuestViewController: BaseViewController {
         )
     }
     
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillHideNotification, object: nil)
-    }
-    
     override func setDelegate() {
         rootView.questTextField.delegate = self
     }
     
-    private func setGesture() {
-        let tapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(endEditingOnTap))
-        let tipTagGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(tipTagDidTap))
-        
-        tapGestureRecognizer.isEnabled = true
-        tapGestureRecognizer.delegate = self
-        tapGestureRecognizer.cancelsTouchesInView = false
-        
-        tipTagGestureRecognizer.isEnabled = true
-        
-        self.rootView.headerView.tipTag.addGestureRecognizer(tipTagGestureRecognizer)
-        self.rootView.scrollView.addGestureRecognizer(tapGestureRecognizer)
-    }
-    
-    private func presentPhotoPicker() {
-        rootView.imageContainer.didTapAddImage = { [weak self] in
-            guard let self = self else { return }
-            self.openPhotosButtenPressed()
-        }
-    }
-}
-
-extension WriteActiveTypeQuestViewController {
     @objc
-    private func textViewMoveUp(_ notification: NSNotification) {
-        guard let keyboardFrame: NSValue = notification.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? NSValue else {
-            return
-        }
-        keyboardFrameInWindow = keyboardFrame.cgRectValue
-        currentKeyboardOffset = 0
-        previousTextViewHeight = rootView.questTextField.textView.bounds.height
-        isKeyboardUsed = true
-        
-        DispatchQueue.main.async { [weak self] in
-            self?.adjustViewForCurrentCaret(animated: true)
-        }
+    override func tipTagDidTap() {
+        let viewController = ViewControllerFactory.shared.makeQuestTipViewController()
+        viewController.bind(questID: questID, questType: questType, questNumber: questNumber)
+        viewController.navigationItem.hidesBackButton = true
+        viewController.modalPresentationStyle = .fullScreen
+        self.present(viewController, animated: false)
     }
     
     @objc
-    private func textViewMoveDown(_ notification: NSNotification) {
-        keyboardFrameInWindow = .zero
-        currentKeyboardOffset = 0
-        previousTextViewHeight = 0
-        UIView.animate(withDuration: 0.3) {
-            self.rootView.transform = .identity
-        }
-        isKeyboardUsed = false
-    }
-    
-    @objc
-    private func confirmButtonDidTap() {
+    override func confirmButtonDidTap() {
         if rootView.questTextField.textView.text == rootView.questTextField.placeholder ||
             rootView.questTextField.textView.text.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
         {
@@ -187,21 +109,15 @@ extension WriteActiveTypeQuestViewController {
             properties: property.dictionary
         )
     }
-    
-    @objc
-    private func endEditingOnTap(sender: UITapGestureRecognizer){
-        self.view.endEditing(true)
-    }
-    
-    @objc
-    private func tipTagDidTap() {
-        let viewController = ViewControllerFactory.shared.makeQuestTipViewController()
-        viewController.bind(questID: questID, questType: questType, questNumber: questNumber)
-        viewController.navigationItem.hidesBackButton = true
-        viewController.modalPresentationStyle = .fullScreen
-        self.present(viewController, animated: false)
+        
+    private func presentPhotoPicker() {
+        rootView.imageContainer.didTapAddImage = { [weak self] in
+            guard let self = self else { return }
+            self.openPhotosButtenPressed()
+        }
     }
 }
+
 
 extension WriteActiveTypeQuestViewController: ToastPresentable, ToastErrorHandler {
     
@@ -298,32 +214,6 @@ extension WriteActiveTypeQuestViewController: ToastPresentable, ToastErrorHandle
                 }
             }
             .store(in: &cancellables)
-    }
-}
-
-extension WriteActiveTypeQuestViewController: BackNavigable {
-    func back() {
-        tabBarController?.tabBar.isHidden = true
-        
-        let action: (() -> Void) = {
-            self.navigationController?.popViewController(animated: true)
-            self.tabBarController?.tabBar.isHidden = false
-        }
-        
-        ModalBuilder(
-            modalView: QuitModalView(),
-            action: action,
-            rootViewController: self
-        ).present()
-    }
-}
-
-extension WriteActiveTypeQuestViewController: UIGestureRecognizerDelegate {
-    func gestureRecognizer(
-        _ agestureRecognizer: UIGestureRecognizer,
-        shouldReceive touch: UITouch)
-    -> Bool {
-        return true
     }
 }
 
@@ -436,65 +326,6 @@ extension WriteActiveTypeQuestViewController: QuestCompleteProtocol {
         viewModel.action(.textFieldEditing(answerText: self.answerText, text: text, imgCount: rootView.imgCount))
         if isKeyboardUsed {
             adjustViewForTextGrowth(animated: false)
-        }
-    }
-}
-
-extension WriteActiveTypeQuestViewController {
-    private func adjustViewForTextGrowth(animated: Bool) {
-        guard isKeyboardUsed else { return }
-        guard !keyboardFrameInWindow.isEmpty else { return }
-        
-        let currentHeight = rootView.questTextField.textView.bounds.height
-        guard currentHeight > 0 else { return }
-        
-        if previousTextViewHeight == 0 {
-            previousTextViewHeight = currentHeight
-            return
-        }
-        
-        let diff = currentHeight - previousTextViewHeight
-        previousTextViewHeight = currentHeight
-        
-        guard abs(diff) > 0.5 else { return }
-
-        let textView = rootView.questTextField.textView
-        let textViewFrameInWindow = textView.convert(textView.bounds, to: nil)
-        let keyboardTop = keyboardFrameInWindow.minY
-        let padding = 12.adjustedH
-
-        let overlap = textViewFrameInWindow.maxY + padding - keyboardTop
-        let targetOffset = max(0, currentKeyboardOffset + overlap)
-
-        guard abs(targetOffset - currentKeyboardOffset) > 0.5 else { return }
-        currentKeyboardOffset = targetOffset
-                
-        UIView.animate(withDuration: 0.2) {
-            self.rootView.transform = CGAffineTransform(translationX: 0, y: -self.currentKeyboardOffset)
-        }
-    }
-    
-    private func adjustViewForCurrentCaret(animated: Bool) {
-        guard isKeyboardUsed else { return }
-        guard !keyboardFrameInWindow.isEmpty else { return }
-        
-        let textView = rootView.questTextField.textView
-        guard textView.isFirstResponder else { return }
-        guard let selectedRange = textView.selectedTextRange else { return }
-        
-        let caretRect = textView.caretRect(for: selectedRange.end)
-        let caretInWindow = textView.convert(caretRect, to: nil)
-        let keyboardTop = keyboardFrameInWindow.minY
-        let padding = 12.adjustedH
-        
-        let overlap = caretInWindow.maxY + padding - keyboardTop
-        let targetOffset = max(0, overlap)
-        
-        guard abs(targetOffset - currentKeyboardOffset) > 0.5 else { return }
-        currentKeyboardOffset = targetOffset
-        
-        UIView.animate(withDuration: 0.2) {
-            self.rootView.transform = CGAffineTransform(translationX: 0, y: -self.currentKeyboardOffset)
         }
     }
 }
